Add webpack-defaults
