+++ .editorconfig
@@ -1,0 +1,13 @@
+# editorconfig.org
+
+[*]
+charset = utf-8
+indent_style = space
+indent_size = 2
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[.md]
+insert_final_newline = false
+trim_trailing_whitespace = false
