+++ .gitattributes
@@ -1,0 +1,3 @@
+package-lock.json -diff
+* text=auto
+bin/* eol=lf
