+++ .travis.yml
@@ -1,6 +1,48 @@
+sudo: false
+dist: trusty
 language: node_js
-matrix:
+branches:
+  only:
+    - master
+jobs:
+  fast_finish: true
+  allow_failures:
+    - env: WEBPACK_VERSION=canary
   include:
-    - node_js: "7"
-    - node_js: "6"
-    - node_js: "4.3"
+    - &test-latest
+      stage: Webpack latest
+      nodejs: 6
+      env: WEBPACK_VERSION=latest JOB_PART=test
+      script: npm run travis:$JOB_PART
+    - <<: *test-latest
+      nodejs: 4.3
+      env: WEBPACK_VERSION=latest JOB_PART=test
+      script: npm run travis:$JOB_PART
+    - <<: *test-latest
+      node_js: 8
+      env: WEBPACK_VERSION=latest JOB_PART=lint
+      script: npm run travis:$JOB_PART
+    - <<: *test-latest
+      node_js: 8
+      env: WEBPACK_VERSION=latest JOB_PART=coverage
+      script: npm run travis:$JOB_PART
+      after_success: 'bash <(curl -s https://codecov.io/bash)'
+    - stage: Webpack canary
+      before_script: npm i --no-save git://github.com/webpack/webpack.git#master
+      script: npm run travis:$JOB_PART
+      node_js: 8
+      env: WEBPACK_VERSION=canary JOB_PART=test
+before_install:
+  - 'if [[ `npm -v` != 5* ]]; then npm i -g npm@^5.0.0; fi'
+  - nvm --version
+  - node --version
+  - npm --version
+before_script:
+  - |-
+    if [ "$WEBPACK_VERSION" ]; then
+      npm i --no-save webpack@$WEBPACK_VERSION
+    fi
+script:
+  - 'npm run travis:$JOB_PART'
+after_success:
+  - 'bash <(curl -s https://codecov.io/bash)'
