+++ LICENSE
@@ -1,4 +1,3 @@
-
 Copyright JS Foundation and other contributors
 
 Permission is hereby granted, free of charge, to any person obtaining
