+++ package.json
@@ -3,50 +3,71 @@
   "version": "1.0.0",
   "author": "Tobias Koppers @sokra",
   "description": "worker loader module for webpack",
-  "scripts": {
-    "test": "mocha",
-    "posttest": "eslint .",
-    "release": "standard-version"
+  "main": "dist/cjs.js",
+  "files": [
+    "dist"
+  ],
+  "engines": {
+    "node": ">= 4.3 < 5.0.0 || >= 5.10"
   },
-  "eslintConfig": {
-    "extends": "webpack",
-    "rules": {
-      "linebreak-style": 0,
-      "comma-dangle": [
-        "error",
-        {
-          "arrays": "always-multiline",
-          "objects": "always-multiline",
-          "imports": "always-multiline",
-          "exports": "always-multiline",
-          "functions": "never"
-        }
-      ],
-      "no-underscore-dangle": 0,
-      "no-param-reassign": 0,
-      "prefer-destructuring": 0,
-      "strict": 0
-    }
+  "scripts": {
+    "test": "jest",
+    "posttest": "eslint .",
+    "webpack-defaults": "webpack-defaults",
+    "start": "npm run build -- -w",
+    "appveyor:test": "npm run test",
+    "build": "cross-env NODE_ENV=production babel src -d dist --ignore 'src/**/*.test.js'",
+    "clean": "del-cli dist",
+    "lint": "eslint --cache src test",
+    "lint-staged": "lint-staged",
+    "prebuild": "npm run clean",
+    "prepublish": "npm run build",
+    "release": "standard-version",
+    "security": "nsp check",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --collectCoverageFrom='src/**/*.js' --coverage",
+    "travis:lint": "npm run lint && npm run security",
+    "travis:test": "npm run test -- --runInBand",
+    "travis:coverage": "npm run test:coverage -- --runInBand"
   },
   "peerDependencies": {
-    "webpack": ">=0.9 || ^2.0.0 || ^3.0.0"
+    "webpack": "^2.0.0 || ^3.0.0"
   },
   "dependencies": {
     "loader-utils": "^1.0.2",
     "schema-utils": "^0.3.0"
   },
   "devDependencies": {
-    "del": "^2.2.2",
+    "babel-cli": "^6.24.1",
+    "babel-jest": "^20.0.3",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-env": "^1.6.0",
+    "cross-env": "^5.0.1",
+    "del": "^3.0.0",
+    "del-cli": "^1.1.0",
     "eslint": "^3.16.0",
     "eslint-config-webpack": "^1.0.0",
     "eslint-plugin-import": "^2.2.0",
+    "jest": "^20.0.4",
+    "lint-staged": "^4.0.0",
     "mocha": "^3.2.0",
+    "nsp": "^2.6.3",
+    "pre-commit": "^1.2.2",
     "standard-version": "^4.2.0",
-    "webpack": "^2.2.1"
+    "webpack": "^2.2.1",
+    "webpack-defaults": "^1.4.0"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/webpack-contrib/worker-loader.git"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "pre-commit": "lint-staged",
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "git add"
+    ]
+  }
 }
